project_manager/requirements.txt
Django==5.0.2
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
flake8==7.0.0
gunicorn==21.2.0
openapi-codec==1.3.2
psycopg2-binary==2.9.9
pytest-django==4.8.0
